pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=65", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "simularium_metrics_calculator"
description = "Calculate plot metrics from spatial agent data"
keywords = []
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache Software License 2.0" }
authors = [
  { email = "blair208@gmail.com", name = "Blair Lyons" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dynamic = ["version"]
dependencies = [
    "simulariumio",
    "scipy",
]

[project.urls]
Homepage = "https://github.com/simularium/metrics-calculator"
"Bug Tracker" = "https://github.com/simularium/metrics-calculator/issues"
Documentation = "https://simularium.github.io/metrics-calculator"
"User Support" = "https://github.com/simularium/metrics-calculator/issues"

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
lint = [
  "black>=22.3.0",
  "check-manifest>=0.48",
  "flake8>=3.8.3",
  "flake8-debugger>=3.2.1",
  "flake8-pyprojecttoml",
  "flake8-typing-imports>=1.9.0",
  "isort>=5.7.0",
  "mypy>=0.790",
  "pre-commit>=2.20.0",
]
test = [
  "coverage>=5.1",
  "pytest>=5.4.3",
  "pytest-cov>=2.9.0",
  "pytest-raises>=0.11",
]
docs = [
  # Sphinx + Doc Gen + Styling
  "m2r2>=0.2.7",
  "Sphinx>=4.0.0",
  "furo>=2022.4.7",
  # Extensions
  "numpydoc",
  "sphinx-copybutton",
  # TODO: Pins
  "docutils>=0.18,<0.19",
]
dev = [
  "ipython>=8.4.0",
]

# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*docs/*", "*tests/*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "py.typed", "*tests/data/*"]

# tools
[tool.black]
line-length = 88

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
profile = "black"
use_parentheses = true

# https://github.com/mgedmin/check-manifest#configuration
[tool.check-manifest]
ignore = [
  ".editorconfig",
  ".pre-commit-config.yaml",
  "CODE_OF_CONDUCT.md",
  "CONTRIBUTING.md",
  "Justfile",
  ".cookiecutter.yaml",
  "*docs/*",
]

[tool.mypy]
files = "simularium_metrics_calculator/*.py"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true

# https://flake8.pycqa.org/en/latest/user/options.html
# https://gitlab.com/durko/flake8-pyprojecttoml
[tool.flake8]
max-line-length = 88
ignore = "E203,E402,W291,W503"
min-python-version = "3.8.0"
per-file-ignores = [
  "__init__.py:F401",
]